restructure tests to use 1 initial config
